+++ src/routes/subscriptions.rs
@@ -1,0 +1,11 @@
+use actix_web::{web, HttpResponse, Responder};
+
+#[derive(serde::Deserialize)]
+pub struct FormData {
+    email: String,
+    name: String,
+}
+
+pub async fn subscribe(form: web::Form<FormData>) -> impl Responder {
+    HttpResponse::Ok()
+}
